PMM-7259 Separate IntegratedAlertingContent
