+++ public/app/features/integrated-alerting/components/IntegratedAlertingContent/IntegratedAlertingContent.constants.ts
