+++ public/app/features/integrated-alerting/components/IntegratedAlertingContent/IntegratedAlertingContent.messages.ts
@@ -1,0 +1,4 @@
+export const Messages = {
+  alertingDisabled: 'Integrated Alerting is disabled. You can enable it in ',
+  pmmSettings: 'PMM Settings.',
+};
