+++ public/app/features/integrated-alerting/components/IntegratedAlertingContent/index.ts
@@ -1,0 +1,1 @@
+export * from './IntegratedAlertingContent';
